# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Codecademy e-Commerce REST API",
      "description": "Defines the backend logic relating to the <strong>'registration'</strong>, <strong>'customers'</strong>, <strong>'addresses'</strong>, <strong>'orders'</strong>, <strong>'products'</strong>, <strong>'cart'</strong>, and <strong>'login'</strong> endpoints of an e-commerce REST API.  <code>GET</code> operations can be performed directly on the <strong>HEROKU</strong> database; <code>POST</code>/<code>PUT</code>/<code>DELETE</code> operations can be performed on the <strong>SWAGGER HUB</strong> <em>virtual mocking server</em> so that the <strong>HEROKU</strong> database is not compromised.",
      "version": "0.1"
    },
    "servers": [
      {
        "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
        "description": "Heroku App"
      },
      {
        "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
        "description": "SwaggerHub API Auto Mocking"
      }
    ],
    "paths": {
...
      "/api/products": {
        "get": {
          "tags": [
            "Products"
          ],
          "summary": "Return all product information",
          "description": "Returns an array of JSON objects that describe the products in the database",
          "responses": {
            "200": {
              "description": "Indicates the successful return of an array of JSON objects that describe the products in the database",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "item_name": "Opportunity Rover - 2008-04-03 (Sol 1490)",
                      "item_description": "Photo taken by the Opportunity Rover's Navigation Camera on 2008-04-03 (expedition sol 1490)",
                      "image_url": "https//mars.nasa.gov/mer/gallery/all/1/n/1490/1N260454912EFF8976P1947L0M1-BR.JPG",
                      "price": 1000
                    },
                    {
                      "id": 2,
                      "item_name": "Opportunity Rover - 2004-02-04 (Sol 10)",
                      "item_description": "Photo taken by the Opportunity Rover's Front Hazard Avoidance Camera on 2004-02-04 (expedition sol 10)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/f/010/1F129070908EDN0224P1101L0M1-BR.JPG",
                      "price": 800
                    },
                    {
                      "id": 3,
                      "item_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                      "item_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2004-01-26 (expedition sol 1)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/001/1P128287426EFF0000P2303L2M1-BR.JPG",
                      "price": 600
                    },
                    {
                      "id": 4,
                      "item_name": "Opportunity Rover - 2009-09-11 (Sol 2001)",
                      "item_description": "Photo taken by the Opportunity Rover's Rear Hazard   Avoidance Camera on 2009-09-11 (expedition sol 2001)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/2001/1R305828935EFFA5FWP1312L0M1-BR.JPG",
                      "price": 700
                    },
                    {
                      "id": 5,
                      "item_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                      "item_description": "Photo taken by the Opportunity Rover's Rear Hazard Avoidance Camera on 2004-01-26 (expedition sol 1)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/001/1R128285184EDN0000P1002L0M1-BR.JPG",
                      "price": 900
                    },
                    {
                      "id": 6,
                      "item_name": "Opportunity Rover - 2006-11-18 (Sol 1001)",
                      "item_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2006-11-18 (expedition sol 1001)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/1001/1P217047759EFF76ZFP2426L2M1-BR.JPG",
                      "price": 1500
                    },
                    {
                      "id": 7,
                      "item_name": "Opportunity Rover - 2009-09-11 (Sol 2001)",
                      "item_description": "Photo taken by the Opportunity Rover's Front Hazard Avoidance Camera on 2009-09-11 (expedition sol 2001)",
                      "image_url": "https://mars.nasa.gov/mer/gallery/all/1/f/2001/1F305828985EFFA5FWP1212L0M1-BR.JPG",
                      "price": 1250
                    },
                    {
                      "id": 8,
                      "item_name": "Item Name [varchar(50)]",
                      "item_description": "Item Description [text]",
                      "image_url": "The NEW URL of the image associated with the item [varchar(100)]",
                      "price": 1000
                    }
                  ]
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "post": {
          "tags": [
            "Products"
          ],
          "summary": "Add a product to the database",
          "description": "Adds a product to the 'products' table of the e-Commerce database",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCreation"
                },
                "examples": {
                  "0": {
                    "value": "{\n    \"item_name\": \"Item Name [varchar(50)]\",\n    \"item_description\": \"Item Description [text]\",\n    \"image_url\": \"The URL of the image associated with the item [varchar(100)]\",\n    \"price\": 1000\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "String response that confirms product creation success",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "example": "Product created successfully."
                }
              }
            },
            "409": {
              "description": "The product's 'image_url' is already in use and thus violates the unique constraint",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "Product creation did not succeed because the image URL is already in use."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
      "/api/products/{productId}": {
        "get": {
          "tags": [
            "Products"
          ],
          "summary": "Return information relating to a particular product",
          "description": "Returns a JSON object that describes the product specified by 'productId'",
          "parameters": [
            {
              "name": "productId",
              "in": "path",
              "description": "The product ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Denotes the successful return of a JSON object that describes the product specified by 'productId'",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  },
                  "example": "{\n  \"id\": 1,\n   \"item_name\": \"Opportunity Rover: 2008-04-03 (Sol 1490)\",\n   \"item_description\": \"Photo taken by the Opportunity Rover's Navigation Camera on 2008-04-03 (expedition sol 1490)\",\n   \"image_url\": \"https//mars.nasa.gov/mer/gallery/all/1/n/1490/1N260454912EFF8976P1947L0M1-BR.JPG\",\n   \"price\": 1000\n}"
                }
              }
            },
            "404": {
              "description": "No product with the specified ID found in the database.",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "No product with the specified ID found in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "put": {
          "tags": [
            "Products"
          ],
          "summary": "Update the details of a particular product",
          "description": "Updates the details of the product specified by 'productId'",
          "parameters": [
            {
              "name": "productId",
              "in": "path",
              "description": "The product ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductUpdate"
                },
                "examples": {
                  "0": {
                    "value": "{\n    \"item_name\": \"Item Name [varchar(50)]\",\n    \"item_description\": \"Item Description [text]\",\n    \"image_url\": \"The UNIQUE URL of the image associated with the item [varchar(100)]\",\n    \"price\": 1250\n}"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "A string that confirms the product update success",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "example": "Product with id {productId} modified successfully."
                }
              }
            },
            "404": {
              "description": "No product with the specified ID found in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "No product with the specified ID found in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "delete": {
          "tags": [
            "Products"
          ],
          "summary": "Delete a product from the database",
          "description": "Auto generated using Swagger Inspector",
          "parameters": [
            {
              "name": "productId",
              "in": "path",
              "description": "The product ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "String that signifies the specified product was successfully deleted",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string"
                  },
                  "example": "Product with id {productId} deleted."
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
...
    },
    "components": {
      "schemas": {
...
        "Product": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "The product ID"
            },
            "item_name": {
              "type": "string",
              "description": "The product name"
            },
            "item_description": {
              "type": "string",
              "description": "The product description"
            },
            "image_url": {
              "type": "string",
              "description": "The URL of the image that represents the item"
            },
            "price": {
              "type": "integer",
              "description": "The price of the product in pence"
            }
          }
        },
        "ProductCreation": {
          "type": "object",
          "properties": {
            "item_name": {
              "type": "string",
              "description": "The product name"
            },
            "item_description": {
              "type": "string",
              "description": "The product description"
            },
            "image_url": {
              "type": "string",
              "description": "The URL of the image that represents the item"
            },
            "price": {
              "type": "integer",
              "description": "The price of the product in pence"
            }
          }
        },
        "ProductUpdate": {
          "type": "object",
          "properties": {
            "item_name": {
              "type": "string",
              "description": "The product name"
            },
            "item_description": {
              "type": "string",
              "description": "The product description"
            },
            "image_url": {
              "type": "string",
              "description": "The URL of the image that represents the product"
            },
            "price": {
              "type": "integer",
              "description": "The price of the product in pence"
            }
          }
        },
...
      }
    }
  }